feat: documented CustomerController and provided hateoas support in all of its endpoints

+++ src/main/java/com/example/demo/services/CustomerService.java
@@ -35,6 +35,8 @@ public class CustomerService {
 	}
 
 	public void deleteById(Long id) {
+		if (!customerRepository.existsById(id))
+			throw new ItemNotFoundException(Customer.class, id);
 		customerRepository.deleteById(id);
 	}
